feat(votos): ahora los votos se guardan en ls

+++ src/js/services/movie.service.js
@@ -17,13 +17,25 @@ class MovieService {
         return new MovieService()
     }
 
+    movieExists(movies, find) {
+        if (!movies) {
+            return false
+        }
+
+        if (Array.isArray(movies)) {
+            return movies.find((movie) => movie.title == find)
+        }
+
+        return movies.title == find
+    }
+
     genreList = new Set()
     addMovie(title, releaseDate, popularity) {
         if (this.genreList.size == 0) {
             alert('Elige un genero')
             return
         }
-
+ 
         const rdDate = new Date(releaseDate)
         const tomorrow = new Date()
         tomorrow.setDate(tomorrow.getDate() + 1)
@@ -37,12 +49,22 @@ class MovieService {
             return
         }
 
+        const resultH2 = document.getElementById('addResult')
         let movies = JSON.parse(localStorage.getItem('movieList'))
+        if (this.movieExists(movies, title)) {
+            resultH2.textContent = 'El titulo de la pelicula ya existe, por favor, elige otro nombre.'
+            resultH2.style.color = '#a50b0b'
+            return
+        }
+
         if (movies && Array.isArray(movies)) {
             movies.push(new MovieEntity(title, releaseDate, popularity, Array.from(this.genreList)))
         } else if (movies) {
             movies = [movies, new MovieEntity(title, releaseDate, popularity)]
         }
+
+        resultH2.textContent = 'Pelicula añadida.'
+        resultH2.style.color = '#52cc29'
         
         localStorage.setItem('movieList', movies ? JSON.stringify(movies) : JSON.stringify(new MovieEntity(title, releaseDate, popularity, Array.from(this.genreList))))
     }
@@ -67,6 +89,31 @@ class MovieService {
         }
         
         this.genreList.add(genre)
+    }
+
+    getMovies() {
+        return JSON.parse(localStorage.getItem('movieList'))
+    }
+
+    vote(vote, id) {
+        const movies = JSON.parse(localStorage.getItem('movieList'))
+        if (Array.isArray(movies)) {
+            const updatedMovies = movies.map((movie) => {
+                if (movie.id == id) {
+                    movie.votes = [...movie.votes, vote]
+                }
+
+                return movie
+            })
+
+            localStorage.setItem('movieList', JSON.stringify(updatedMovies))
+            return
+        }
+
+        if (movies.id == id) {
+            movies.votes = [...movies.votes, vote]
+            localStorage.setItem('movieList', movies)
+        }
     }
 }
